pyproject.toml
[project]
name = "rag-model-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=2.2.2",
    "groq>=0.1.0",
    "httpx>=0.24.0",
]
